// playground-react/src/App/Pages/Reveals/LensPage/LensPage.types.ts
import type { Point2d, Size2d } from "@thewaver/ss-utils";

export type LensExampleProps = {
    zoom: number;
    radius: number;
    softness: number;
    stepSize: number;
    joinRadii: number[];
    lameExponents: number[];
    isDisabled: boolean;
    computePoints: ((size: Size2d) => Point2d[]) | undefined;
};